feat(createWatcher): add creating watcher route
