+++ lib/create_watcher.dart
@@ -1,0 +1,40 @@
+import 'package:flutter/material.dart';
+
+class CreateWatcherRoute extends StatefulWidget {
+  const CreateWatcherRoute({Key? key, required this.title}) : super(key: key);
+
+  final String title;
+
+  @override
+  State<CreateWatcherRoute> createState() => _CreateWatcherRouteState();
+}
+
+class _CreateWatcherRouteState extends State<CreateWatcherRoute> {
+  void _noop() {}
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        backgroundColor: Theme.of(context).colorScheme.inversePrimary,
+        title: Text(widget.title),
+      ),
+      body: Center(
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: <Widget>[
+            Text(
+              'Hello, World!',
+              style: Theme.of(context).textTheme.titleLarge,
+            ),
+          ],
+        ),
+      ),
+      floatingActionButton: FloatingActionButton(
+        onPressed: _noop,
+        tooltip: 'Increment',
+        child: const Icon(Icons.add),
+      ),
+    );
+  }
+}
