+++ lib/home.dart
@@ -1,0 +1,49 @@
+import 'package:flutter/material.dart';
+
+import 'create_watcher.dart';
+
+class HomeRoute extends StatefulWidget {
+  const HomeRoute({super.key, required this.title});
+
+  final String title;
+
+  @override
+  State<HomeRoute> createState() => _HomeRouteState();
+}
+
+class _HomeRouteState extends State<HomeRoute> {
+  void handleAddNewWatcherClick() {
+    Navigator.push(
+      context,
+      MaterialPageRoute(
+        builder: (context) => const CreateWatcherRoute(title: 'Create Watcher'),
+      ),
+    );
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        backgroundColor: Theme.of(context).colorScheme.inversePrimary,
+        title: Text(widget.title),
+      ),
+      body: Center(
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: <Widget>[
+            Text(
+              'Hello, World!',
+              style: Theme.of(context).textTheme.titleLarge,
+            ),
+          ],
+        ),
+      ),
+      floatingActionButton: FloatingActionButton(
+        onPressed: handleAddNewWatcherClick,
+        tooltip: 'Increment',
+        child: const Icon(Icons.add),
+      ),
+    );
+  }
+}
